Put PermutationVmn into vmn package. Clarify what Permutation does.
