+++ src/main/kotlin/org/cryptobiotic/mixnet/core/Permutation.kt
@@ -4,10 +4,14 @@ import java.security.SecureRandom
 
 /**
     psi: {0..N-1} -> {0..N-1}
-    for vector e, pe = psi.permute(e). then ei = pej, where j = psi.inv(i), i = psi.of(j)
+    for any vector e, pe = psi.permute(e). then e_i = pe_j, where j = psi.inv(i), i = psi.of(j)
     so e[idx] = pe[psi.inv(idx)]; you have an element in e, and need to get the corresponding element from pe
     so pe[jdx] = e[ps.of(jdx)]; you have an element in pe, and need to get the corresponding element from e
-*/
+
+   (ie)  <- invert (e) permute -> (pe)
+   (ie) permute -> (e)  <- invert (pe)
+
+ */
 
 data class Permutation(private val psi: IntArray) {
     val n = psi.size
@@ -69,83 +73,6 @@ data class PermutationMatrix( val elems: List<IntArray> ) {
         return buildString {
             elems.forEach { append("${it.contentToString()}\n")
             }
-        }
-    }
-}
-
-// port from vmn code, for testing
-data class VmnPermutation(private val table: IntArray) {
-    val n = table.size
-
-    fun <T> applyPermutation(
-        array: List<T>,
-        permutedArray: MutableList<T>
-    ) {
-        for (i in array.indices) {
-            permutedArray[table[i]] = array[i]
-        }
-    }
-
-    fun inverse(): VmnPermutation {
-        val invtable = IntArray(n)
-        for (i in table.indices) {
-            invtable[table[i]] = i
-        }
-        return VmnPermutation(invtable)
-    }
-
-    companion object {
-        fun random(n: Int) : VmnPermutation {
-            val result = MutableList(n) { it }
-            result.shuffle(SecureRandom.getInstanceStrong())
-            return VmnPermutation(result.toIntArray())
-        }
-    }
-}
-
-//////////////////////////////////////////////////////////////
-// move to vmn
-
-//     public void applyPermutation(final Object[] array,
-//                                 final Object[] permutedArray) {
-//        final LargeInteger[] integers = table.integers();
-//        for (int i = 0; i < array.length; i++) {
-//            permutedArray[integers[i].intValue()] = array[i];
-//        }
-//    }
-
-//     public Permutation inv() {
-//        final LargeInteger[] invtable = new LargeInteger[table.size()];
-//        final LargeInteger[] orig = table.integers();
-//
-//        for (int i = 0; i < orig.length; i++) {
-//            invtable[orig[i].intValue()] = new LargeInteger(i);
-//        }
-//        return new PermutationIM(new LargeIntegerArrayIM(invtable));
-//    }
-
-// matches Vmn PermutationIM
-data class PermutationVmn(private val inverse: IntArray) {
-    val n = inverse.size
-    private val table: IntArray
-    init {
-        table = IntArray(n)
-        for (idx in inverse) {
-            table[inverse[idx]] = idx
-        }
-    }
-
-    fun inv(idx:Int) = inverse[idx]
-    fun of(jdx:Int) = table[jdx]
-
-    fun <T> invert(list: List<T>): List<T> = List(list.size) { idx -> list[inverse[idx]] }
-    fun <T> permute(list: List<T>): List<T> = List(list.size) { idx -> list[table[idx]] }
-
-    companion object {
-        fun random(n: Int) : PermutationVmn {
-            val result = MutableList(n) { it }
-            result.shuffle(SecureRandom.getInstanceStrong())
-            return PermutationVmn(result.toIntArray())
         }
     }
 }
